home-media-app/src/main/java/com/gorolykmaxim/homemediaapp/service/view/tvshow/TvShowPrototype.java
package com.gorolykmaxim.homemediaapp.service.view.tvshow;

import java.util.Objects;

public class TvShowPrototype {
    private String name;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        TvShowPrototype that = (TvShowPrototype) o;
        return Objects.equals(name, that.name);
    }

    @Override
    public int hashCode() {
        return Objects.hash(name);
    }
}
